Adds thumbnail to show page.
