+++ app/assets/stylesheets/search.scss
@@ -33,12 +33,3 @@ a[aria-expanded="false"] {
 .accordion-header.facet-field-heading .accordion-button:disabled::after {
   display: none;
 }
-
-.thumbnail {
-  max-width: 240px;
-  max-height: 240px;
-  border-color: var(--stanford-30-black);
-  border-width: 1px;
-  border-radius: 4px;
-  border-style: solid;
-}
