+++ app/assets/stylesheets/show.scss
@@ -1,0 +1,20 @@
+// Styles for the show page
+
+.nav-underline {
+  border-bottom: 1px solid black;
+  margin-bottom: 10px;
+
+  .nav-link {
+    width: 10.5rem;
+  }
+
+  .nav-link:not(.active, .disabled) {
+    color: rgb(var(--bs-link-color-rgb), var(--bs-link-opacity, 1));
+  }
+
+  .nav-link.active {
+    border-bottom: 3px solid #B1040E;
+    color: #B1040E;
+    font-weight: bold;
+  }
+}
