+++ app/components/search/item_result_component.html.erb
@@ -31,7 +31,7 @@
       <% end %>
     </div>
     <div class="col-md-2 mx-auto align-self-center">
-      <%= render Search::ThumbnailComponent.new(result:) %>
+      <%= render ThumbnailComponent.new(result:) %>
     </div>
   </div>
 <% end %>
