+++ app/components/thumbnail_component.html.erb
